requirements.txt
aniso8601==7.0.0
asgiref==3.5.0
astroid==2.9.3
attrs==21.4.0
autopep8==1.6.0
certifi==2020.6.20
charset-normalizer==2.0.12
coreapi==2.3.3
coreschema==0.0.4
Django==4.0.2
django-ckeditor==6.2.0
django-cors-headers==3.11.0
django-filter==21.1
django-js-asset==2.0.0
djangorestframework==3.13.1
djangorestframework-camel-case==1.3.0
djangorestframework-simplejwt==5.1.0
drf-yasg==1.20.0
Faker==12.0.1
flake8==4.0.1
graphene==2.1.9
graphene-django==2.15.0
graphql-core==2.3.2
graphql-relay==2.0.1
idna==3.3
inflection==0.5.1
iniconfig==1.1.1
isort==5.10.1
itypes==1.2.0
jedi==0.18.1
Jinja2==3.1.1
lazy-object-proxy==1.7.1
MarkupSafe==2.1.1
mccabe==0.6.1
mixer==7.2.2
packaging==21.3
parso==0.8.3
Pillow==9.0.1
platformdirs==2.5.1
pluggy==1.0.0
promise==2.3
psycopg2==2.8.6
psycopg2-binary==2.9.3
py==1.11.0
pycodestyle==2.8.0
pydocstyle==6.1.1
pyflakes==2.4.0
PyJWT==2.3.0
pylint==2.12.2
pyparsing==3.0.7
pytest==7.1.1
python-dateutil==2.8.2
python-lsp-jsonrpc==1.0.0
python-lsp-server==1.4.0
pytz==2021.3
requests==2.27.1
rope==0.23.0
ruamel.yaml==0.17.21
ruamel.yaml.clib==0.2.6
Rx==1.6.1
singledispatch==3.7.0
six==1.16.0
snowballstemmer==2.2.0
sqlparse==0.4.2
text-unidecode==1.3
toml==0.10.2
tomli==2.0.1
ujson==5.1.0
uritemplate==4.1.1
urllib3==1.26.9
wrapt==1.13.3
yapf==0.32.0
